add custom field test, remove references in the card model to keep this branch clean, re #7842
