+++ arches/app/models/graph.py
@@ -29,7 +29,7 @@ from arches.app.models import models
 from arches.app.models.resource import Resource
 from arches.app.models.system_settings import settings
 from arches.app.datatypes.datatypes import DataTypeFactory
-from arches.app.utils.language import default_lang_node_json
+from arches.app.models.fields.i18n import default_lang_node_json
 from arches.app.utils.betterJSONSerializer import JSONSerializer, JSONDeserializer
 from arches.app.search.search_engine_factory import SearchEngineFactory
 from django.utils.translation import ugettext as _
