+++ arches/app/models/models.py
@@ -15,7 +15,6 @@ import uuid
 import datetime
 import logging
 from datetime import timedelta
-from arches.app.utils.language import default_lang_node_json
 from arches.app.utils.module_importer import get_class_from_modulename
 from django.forms.models import model_to_dict
 from django.contrib.gis.db import models
@@ -40,86 +39,9 @@ from guardian.shortcuts import assign_perm
 from django.conf import settings
 
 
-
-from arches.app.models.system_settings import settings
-from django.utils.translation import get_language
-from django.contrib.postgres.fields.jsonb import JsonAdapter
-class JSONBSet(object):
-    def __init__(self, attname, value):
-        self.sql = "jsonb_set(" + attname + ", %s, %s)"
-        self.value = value
-
-    def as_sql(self, compiler, connection):
-        lang = f"{{{get_language()}}}"
-        params = (lang, json.dumps(self.value))
-        return self.sql, params
-
-class L10n_Field(object):
-    def __init__(self, value):
-        self.value = value
-
-    def __str__(self):
-        if self.value is None:
-            return self.value
-
-        try:
-            ret = json.loads(self.value)
-        except TypeError:
-            ret = self.value
-
-        try:
-            return ret[get_language()]
-        except KeyError as e:
-            try:
-                return ret[settings.LANGUAGE_CODE]
-            except KeyError as e:
-                try:
-                    return list(ret.values())[0]
-                except:
-                    print('error')
-                    return ""
-
-    def serialize(self):
-        return str(self)
-
-
-class TranlatedJSONField(JSONField):
-    def from_db_value(self, value, expression, connection):
-        # import ipdb; ipdb.sset_trace()
-        if value is not None:
-            return L10n_Field(value)
-        return None
-
-    def to_python(self, value):
-        # import ipdb; ipdb.sset_trace()
-        if isinstance(value, L10n_Field):
-            return value
-        if value is None:
-            return value
-        value = super().to_python(value)
-        return L10n_Field(value)
-
-    def get_prep_value(self, value):
-        # import ipdb; ipdb.sset_trace()
-        # print(f'in get_prep_value, value={value}')
-        if isinstance(value, str):
-            try:
-                json.loads(value)
-            except:
-                value = JSONBSet(self.attname, value)
-        elif isinstance(value, L10n_Field):
-            value = json.dumps(value.value)
-        elif isinstance(value, dict):
-            value = json.dumps(value)
-        return value
-
-    def get_db_prep_value(self, value, connection, prepared=False):
-        # print(f'in get_db_prep_value, value={value}')
-        return super().get_db_prep_value(value, connection, prepared)
-
 class CardModel(models.Model):
     cardid = models.UUIDField(primary_key=True, default=uuid.uuid1)  # This field type is a guess.
-    name = TranlatedJSONField(blank=True, null=True, default=default_lang_node_json)
+    name = models.TextField(blank=True, null=True)
     description = models.TextField(blank=True, null=True)
     instructions = models.TextField(blank=True, null=True)
     cssclass = models.TextField(blank=True, null=True)
